refactor(suite-build): extract shared vite config and browser polyfills
moved the workspace aliases, the core-js/actual no-op plugin and the browser
polyfills out of vite.config.mts into viteShared.ts and browserPolyfills.ts so
a second vite config can reuse them. no behaviour change.
